third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/tentative/restrict-properties/iframe-popup.https_9-last-expected.txt

This is a testharness.js-based test.
[FAIL] COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, cross origin popup with window_open
  assert_equals: Popup name is cleared? expected "true" but got "false"
[FAIL] COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, cross origin popup with anchor
  assert_equals: Popup name is cleared? expected "true" but got "false"
[FAIL] COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, cross origin popup with form
  assert_equals: Popup name is cleared? expected "true" but got "false"
Harness: the test ran to completion.
